lesson:1 d

+++ Занятие 1/Практические задания/!ProgramTasks/D/Program.cs
@@ -6,8 +6,8 @@
         {
             double amount = 1.11; //количество биткоинов от одного человека
             int peopleCount = 60; // количество человек
-            int totalMoney = (int)amount * (int)peopleCount; // ← исправьте ошибку в этой строке
-            Console.WriteLine(totalMoney);
+            double totalMoney  = (int)amount * (int)peopleCount; // ← исправьте ошибку в этой строке
+            Console.WriteLine((int)totalMoney);
         }
     }
 }
